Move the enum to a separate module
