+++ lib/spack/spack/spec.py
@@ -95,6 +95,8 @@ import spack.variant as vt
 import spack.version as vn
 import spack.version.git_ref_lookup
 
+from .enums import ANY_STATUS
+
 __all__ = [
     "CompilerSpec",
     "Spec",
@@ -2066,7 +2068,6 @@ class Spec:
         """Lookup just one spec with an abstract hash, returning a spec from the the environment,
         store, or finally, binary caches."""
         import spack.binary_distribution
-        import spack.database as db
         import spack.environment
 
         active_env = spack.environment.active_environment()
@@ -2074,7 +2075,7 @@ class Spec:
         # First env, then store, then binary cache
         matches = (
             (active_env.all_matching_specs(self) if active_env else [])
-            or spack.store.STORE.db.query(self, installed=db.ANY_STATUS)
+            or spack.store.STORE.db.query(self, installed=ANY_STATUS)
             or spack.binary_distribution.BinaryCacheQuery(True)(self)
         )
